# docs/swagger.yaml
basePath: /api
definitions:
  handler.CoderUrlRequest:
    properties:
      url:
        type: string
    required:
    - url
    type: object
  handler.ErrorResponse:
    properties:
      errorCode:
        example: "400"
        type: string
      message:
        example: error on create
        type: string
    type: object
  handler.SuccessResponse:
    properties:
      code:
        example: "201"
        type: string
      data: {}
    type: object
info:
  contact: {}
  description: api to store and shorten urls
  title: Shortener url go api
  version: "1.0"
paths:
  /url/codeUrl:
    post:
      consumes:
      - application/json
      description: Accepts an original URL as a query parameter and returns a shortened
        URL.
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CoderUrlRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Shortened URL created successfully
          schema:
            $ref: '#/definitions/handler.SuccessResponse'
        "400":
          description: Invalid parameter or URL already shortened
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal error while processing the URL
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Shorten URL
      tags:
      - URL
  /url/decodeUrl/{hash}:
    get:
      consumes:
      - application/json
      description: Takes a shortened URL and returns the original URL.
      parameters:
      - description: Shortened URL to decode
        in: path
        name: hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Original URL
          schema:
            $ref: '#/definitions/handler.SuccessResponse'
        "400":
          description: Error URL not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal error while searching the URL
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Decode URL
      tags:
      - URL
swagger: "2.0"